requirements.txt
appdirs==1.4.4
asgiref==3.3.1
blis==0.2.4
cassandra-driver==3.25.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cymem==2.0.2
decorator==4.4.0
distlib==0.3.1
en-core-web-sm==2.1.0
filelock==3.0.12
Flask==1.1.1
Flask-Caching==1.9.0
Flask-Cors==3.0.10
future==0.17.1
geomet==0.2.1.post1
gunicorn==19.9.0
idna==2.8
importlib-metadata==3.1.1
importlib-resources==3.3.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
murmurhash==1.0.2
networkx==2.3
nltk==3.4.5
numpy==1.17.0
pandas==0.25.0
pke==1.8.1
plac==0.9.6
preshed==2.0.1
python-dateutil==2.8.0
pytz==2019.2
rake-nltk==1.0.4
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
spacy==2.1.8
sqlparse==0.4.1
srsly==0.1.0
thinc==7.0.8
tqdm==4.34.0
Unidecode==1.1.1
urllib3==1.25.3
virtualenv==20.2.1
wasabi==0.2.2
Werkzeug==0.15.5
zipp==3.4.0
